# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.28.0"
  },
  "paths": {
...
    "/reply": {
      "post": {
        "tags": [
          "super::routes::reply"
        ],
        "operationId": "reply",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Streaming response initiated",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/MessageEvent"
                }
              }
            }
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Session Management"
        ],
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "type": "object",
        "required": [
          "user_message",
          "session_id"
        ],
        "properties": {
          "override_conversation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "Override the server's conversation history. Only use this when you need absolute control\nover the conversation state (e.g., administrative tools). For normal operations, the server\nis the source of truth - use truncate/fork endpoints to modify conversation history instead.",
            "nullable": true
          },
          "recipe_name": {
            "type": "string",
            "nullable": true
          },
          "recipe_version": {
            "type": "string",
            "nullable": true
          },
          "session_id": {
            "type": "string"
          },
          "user_message": {
            "$ref": "#/components/schemas/Message"
          }
        }
      },
...
      "Conversation": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Message"
        }
      },
...
      "ExtensionData": {
        "type": "object",
        "description": "Extension data containing all extension states\nKeys are in format \"extension_name.version\" (e.g., \"todo.v0\")",
        "additionalProperties": {}
      },
...
      "GooseMode": {
        "type": "string",
        "enum": [
          "auto",
          "approve",
          "smart_approve",
          "chat"
        ]
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "created",
          "content",
          "metadata"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "MessageEvent": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "message",
              "token_state",
              "type"
            ],
            "properties": {
              "message": {
                "$ref": "#/components/schemas/Message"
              },
              "token_state": {
                "$ref": "#/components/schemas/TokenState"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Message"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "error",
              "type"
            ],
            "properties": {
              "error": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Error"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "reason",
              "token_state",
              "type"
            ],
            "properties": {
              "reason": {
                "type": "string"
              },
              "token_state": {
                "$ref": "#/components/schemas/TokenState"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Finish"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "request_id",
              "message",
              "type"
            ],
            "properties": {
              "message": {
                "type": "object"
              },
              "request_id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Notification"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "conversation",
              "type"
            ],
            "properties": {
              "conversation": {
                "$ref": "#/components/schemas/Conversation"
              },
              "type": {
                "type": "string",
                "enum": [
                  "UpdateConversation"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Sent at the start of an SSE stream to inform the client about\nin-flight requests it can reattach to.",
            "required": [
              "request_ids",
              "type"
            ],
            "properties": {
              "request_ids": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "ActiveRequests"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Ping"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
      "MessageMetadata": {
        "type": "object",
        "description": "Metadata for message visibility",
        "required": [
          "userVisible",
          "agentVisible"
        ],
        "properties": {
          "agentVisible": {
            "type": "boolean",
            "description": "Whether the message should be included in the agent's context window"
          },
          "userVisible": {
            "type": "boolean",
            "description": "Whether the message should be visible to the user in the UI"
          }
        }
      },
...
      "ModelConfig": {
        "type": "object",
        "required": [
          "model_name",
          "toolshim"
        ],
        "properties": {
          "context_limit": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "max_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "model_name": {
            "type": "string"
          },
          "reasoning": {
            "type": "boolean",
            "nullable": true
          },
          "request_params": {
            "type": "object",
            "description": "Provider-specific request parameters (e.g., anthropic_beta headers)",
            "additionalProperties": {},
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "toolshim": {
            "type": "boolean"
          },
          "toolshim_model": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Recipe": {
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "properties": {
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Author"
              }
            ],
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "extensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionConfig"
            },
            "nullable": true
          },
          "instructions": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeParameter"
            },
            "nullable": true
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Response"
              }
            ],
            "nullable": true
          },
          "retry": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RetryConfig"
              }
            ],
            "nullable": true
          },
          "settings": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Settings"
              }
            ],
            "nullable": true
          },
          "sub_recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubRecipe"
            },
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "Role": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "user"
            ]
          },
          {
            "type": "string",
            "enum": [
              "assistant"
            ]
          }
        ]
      },
...
      "Session": {
        "type": "object",
        "required": [
          "id",
          "working_dir",
          "name",
          "created_at",
          "updated_at",
          "extension_data",
          "message_count"
        ],
        "properties": {
          "accumulated_cache_read_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_cache_write_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_output_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_total_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cache_read_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cache_write_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "conversation": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Conversation"
              }
            ],
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "extension_data": {
            "$ref": "#/components/schemas/ExtensionData"
          },
          "goose_mode": {
            "$ref": "#/components/schemas/GooseMode"
          },
          "id": {
            "type": "string"
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "message_count": {
            "type": "integer",
            "minimum": 0
          },
          "model_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ModelConfig"
              }
            ],
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "provider_name": {
            "type": "string",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          },
          "resolved_model_name": {
            "type": "string",
            "nullable": true
          },
          "schedule_id": {
            "type": "string",
            "nullable": true
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_recipe_values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "user_set_name": {
            "type": "boolean"
          },
          "working_dir": {
            "type": "string"
          }
        }
      },
...
      "SessionType": {
        "type": "string",
        "enum": [
          "user",
          "scheduled",
          "sub_agent",
          "hidden",
          "terminal",
          "gateway"
        ]
      },
...
      "TokenState": {
        "type": "object",
        "required": [
          "inputTokens",
          "outputTokens",
          "totalTokens",
          "cacheReadInputTokens",
          "cacheWriteInputTokens",
          "accumulatedInputTokens",
          "accumulatedOutputTokens",
          "accumulatedTotalTokens",
          "accumulatedCacheReadInputTokens",
          "accumulatedCacheWriteInputTokens"
        ],
        "properties": {
          "accumulatedCacheReadInputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "accumulatedCacheWriteInputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "accumulatedInputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "accumulatedOutputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "accumulatedTotalTokens": {
            "type": "integer",
            "format": "int32"
          },
          "cacheReadInputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "cacheWriteInputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "inputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "outputTokens": {
            "type": "integer",
            "format": "int32"
          },
          "totalTokens": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}